server: Split metrics and health code
